requirements.txt
Djangorestframework==3.0
beautifulsoup4==4.3.2
celery==3.1.17
django-celery==3.1.16
django==1.7.5
django-filter=0.9.2
ipython==3.0.0
lxml==3.4.2
numpy==1.9.1
psycopg2==2.6
pyyaml==3.11
redis==2.10.3
requests==2.5.3
scrapy==0.24.5
selenium==2.45.0
service_identity==14.0.0
